Move MSA class into script file
